Backend/e_santebackend/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cobble==0.1.4
colorama==0.4.6
coloredlogs==15.0.1
cryptography==44.0.2
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
et_xmlfile==2.0.0
filelock==3.16.1
pdfminer.six==20250327
pillow==11.1.0
pipenv==2024.4.0
platformdirs==4.3.6
protobuf==6.30.2
psycopg2==2.9.10
psycopg2-binary==2.9.9
pycparser==2.22
pydub==0.25.1
PyJWT==2.9.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
qrcode==8.1
reportlab==4.4.0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.6.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
